% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_legend_colored_exprs.R
\name{mod_plots_legend_colored_exprs_ui}
\alias{mod_plots_legend_colored_exprs_ui}
\alias{mod_plots_legend_colored_exprs_server}
\title{mod_legend_colored_exprs_ui and mod_legend_colored_exprs_server}
\usage{
mod_plots_legend_colored_exprs_ui(id)

mod_plots_legend_colored_exprs_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
