% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_corr_matrix.R
\name{mod_plots_corr_matrix_ui}
\alias{mod_plots_corr_matrix_ui}
\alias{mod_plots_corr_matrix_server}
\title{mod_plots_corr_matrix_ui and mod_plots_corr_matrix_server}
\usage{
mod_plots_corr_matrix_ui(id)

mod_plots_corr_matrix_server(id, obj, names = NULL, gradientRate = NULL)
}
\arguments{
\item{id}{shiny id}

\item{obj}{xxx}

\item{names}{xxx}

\item{gradientRate}{xxx. Default value is 0.9}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
