% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_density.R
\name{mod_plots_density_ui}
\alias{mod_plots_density_ui}
\alias{mod_plots_density_server}
\title{mod_plots_density_ui and mod_plots_density_server}
\usage{
mod_plots_density_ui(id)

mod_plots_density_server(id, obj, conds, legend = NULL, base_palette = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
