% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_group_mv.R
\name{mod_plots_group_mv_ui}
\alias{mod_plots_group_mv_ui}
\alias{mod_plots_group_mv_server}
\title{mod_plots_group_mv_ui and mod_plots_group_mv_server}
\usage{
mod_plots_group_mv_ui(id)

mod_plots_group_mv_server(id, obj, conds, base_palette)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}

\item{conds}{xxx}

\item{base_palette}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
