% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_heatmap.R
\name{mod_plots_heatmap_ui}
\alias{mod_plots_heatmap_ui}
\alias{mod_plots_heatmap_server}
\title{mod_plots_heatmap_ui and mod_plots_heatmap_server}
\usage{
mod_plots_heatmap_ui(id)

mod_plots_heatmap_server(id, obj, conds, width = 900)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
