% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_mv_for_imputation.R
\name{mod_plots_mv_for_imputation_ui}
\alias{mod_plots_mv_for_imputation_ui}
\alias{mod_plots_mv_for_imputation_server}
\title{mod_plots_mv_ui and mod_plots_mv_server}
\usage{
mod_plots_mv_for_imputation_ui(id)

mod_plots_mv_for_imputation_server(id, obj, ind, title = NULL, palette = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
