% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_var_dist.R
\name{mod_plots_var_dist_ui}
\alias{mod_plots_var_dist_ui}
\alias{mod_plots_var_dist_server}
\title{mod_var_dist_plot_ui and mod_var_dist_plot_server}
\usage{
mod_plots_var_dist_ui(id)

mod_plots_var_dist_server(id, obj, conds, base_palette = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
