% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peptidomic.R
\name{Peptidomic}
\alias{Peptidomic}
\title{Simple pipeline for test purpose}
\description{
This pipeline inherits from the class `Pipeline` and have only one private field `.config`.
}
\section{Super classes}{
\code{\link[Magellan:ScreenManager]{Magellan::ScreenManager}} -> \code{\link[Magellan:Pipeline]{Magellan::Pipeline}} -> \code{Peptidomic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{Peptidomic$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Change_Current_Pos">}\href{../../Magellan/html/ScreenManager.html#method-Change_Current_Pos}{\code{Magellan::ScreenManager$Change_Current_Pos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Get_Result">}\href{../../Magellan/html/ScreenManager.html#method-Get_Result}{\code{Magellan::ScreenManager$Get_Result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Reseted">}\href{../../Magellan/html/ScreenManager.html#method-Set_Reseted}{\code{Magellan::ScreenManager$Set_Reseted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Skipped">}\href{../../Magellan/html/ScreenManager.html#method-Set_Skipped}{\code{Magellan::ScreenManager$Set_Skipped()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ValidatePosition">}\href{../../Magellan/html/ScreenManager.html#method-ValidatePosition}{\code{Magellan::ScreenManager$ValidatePosition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="initialize">}\href{../../Magellan/html/ScreenManager.html#method-initialize}{\code{Magellan::ScreenManager$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="server">}\href{../../Magellan/html/ScreenManager.html#method-server}{\code{Magellan::ScreenManager$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ui">}\href{../../Magellan/html/ScreenManager.html#method-ui}{\code{Magellan::ScreenManager$ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="ActionOn_Data_Trigger">}\href{../../Magellan/html/Pipeline.html#method-ActionOn_Data_Trigger}{\code{Magellan::Pipeline$ActionOn_Data_Trigger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="ActionOn_NewPosition">}\href{../../Magellan/html/Pipeline.html#method-ActionOn_NewPosition}{\code{Magellan::Pipeline$ActionOn_NewPosition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="ActionOn_New_DataIn">}\href{../../Magellan/html/Pipeline.html#method-ActionOn_New_DataIn}{\code{Magellan::Pipeline$ActionOn_New_DataIn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="Additional_Server_Funcs">}\href{../../Magellan/html/Pipeline.html#method-Additional_Server_Funcs}{\code{Magellan::Pipeline$Additional_Server_Funcs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="Discover_Skipped_Steps">}\href{../../Magellan/html/Pipeline.html#method-Discover_Skipped_Steps}{\code{Magellan::Pipeline$Discover_Skipped_Steps()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="EncapsulateScreens">}\href{../../Magellan/html/Pipeline.html#method-EncapsulateScreens}{\code{Magellan::Pipeline$EncapsulateScreens()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="GetMaxValidated_BeforeCurrentPos">}\href{../../Magellan/html/Pipeline.html#method-GetMaxValidated_BeforeCurrentPos}{\code{Magellan::Pipeline$GetMaxValidated_BeforeCurrentPos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="GetMaxValidated_BeforePos">}\href{../../Magellan/html/Pipeline.html#method-GetMaxValidated_BeforePos}{\code{Magellan::Pipeline$GetMaxValidated_BeforePos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="GetScreens_ui">}\href{../../Magellan/html/Pipeline.html#method-GetScreens_ui}{\code{Magellan::Pipeline$GetScreens_ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="Global_server">}\href{../../Magellan/html/Pipeline.html#method-Global_server}{\code{Magellan::Pipeline$Global_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="Launch_Module_Server">}\href{../../Magellan/html/Pipeline.html#method-Launch_Module_Server}{\code{Magellan::Pipeline$Launch_Module_Server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="PrepareData2Send">}\href{../../Magellan/html/Pipeline.html#method-PrepareData2Send}{\code{Magellan::Pipeline$PrepareData2Send()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="Set_All_Reset">}\href{../../Magellan/html/Pipeline.html#method-Set_All_Reset}{\code{Magellan::Pipeline$Set_All_Reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="ToggleState_Screens">}\href{../../Magellan/html/Pipeline.html#method-ToggleState_Screens}{\code{Magellan::Pipeline$ToggleState_Screens()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Pipeline" data-id="ValidateCurrentPos">}\href{../../Magellan/html/Pipeline.html#method-ValidateCurrentPos}{\code{Magellan::Pipeline$ValidateCurrentPos()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Peptidomic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
