% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_Normalization.R
\name{Protein_Normalization}
\alias{Protein_Normalization}
\title{Reactive behavior : Normalization of data}
\description{
Reactive behavior : Normalization of data

Reactive behavior : Normalization of data
}
\author{
Samuel Wieczorek
}
\section{Super classes}{
\code{\link[Magellan:ScreenManager]{Magellan::ScreenManager}} -> \code{\link[Magellan:Process]{Magellan::Process}} -> \code{Protein_Normalization}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Global_server}{\code{Protein_Normalization$Global_server()}}
\item \href{#method-Description_server}{\code{Protein_Normalization$Description_server()}}
\item \href{#method-Description_ui}{\code{Protein_Normalization$Description_ui()}}
\item \href{#method-Normalize_server}{\code{Protein_Normalization$Normalize_server()}}
\item \href{#method-Normalize_ui}{\code{Protein_Normalization$Normalize_ui()}}
\item \href{#method-Save_server}{\code{Protein_Normalization$Save_server()}}
\item \href{#method-Save_ui}{\code{Protein_Normalization$Save_ui()}}
\item \href{#method-clone}{\code{Protein_Normalization$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ActionOn_NewPosition">}\href{../../Magellan/html/ScreenManager.html#method-ActionOn_NewPosition}{\code{Magellan::ScreenManager$ActionOn_NewPosition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ActionOn_New_DataIn">}\href{../../Magellan/html/ScreenManager.html#method-ActionOn_New_DataIn}{\code{Magellan::ScreenManager$ActionOn_New_DataIn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Additional_Server_Funcs">}\href{../../Magellan/html/ScreenManager.html#method-Additional_Server_Funcs}{\code{Magellan::ScreenManager$Additional_Server_Funcs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Change_Current_Pos">}\href{../../Magellan/html/ScreenManager.html#method-Change_Current_Pos}{\code{Magellan::ScreenManager$Change_Current_Pos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Get_Result">}\href{../../Magellan/html/ScreenManager.html#method-Get_Result}{\code{Magellan::ScreenManager$Get_Result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Reseted">}\href{../../Magellan/html/ScreenManager.html#method-Set_Reseted}{\code{Magellan::ScreenManager$Set_Reseted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Skipped">}\href{../../Magellan/html/ScreenManager.html#method-Set_Skipped}{\code{Magellan::ScreenManager$Set_Skipped()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ValidatePosition">}\href{../../Magellan/html/ScreenManager.html#method-ValidatePosition}{\code{Magellan::ScreenManager$ValidatePosition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="initialize">}\href{../../Magellan/html/ScreenManager.html#method-initialize}{\code{Magellan::ScreenManager$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="server">}\href{../../Magellan/html/ScreenManager.html#method-server}{\code{Magellan::ScreenManager$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ui">}\href{../../Magellan/html/ScreenManager.html#method-ui}{\code{Magellan::ScreenManager$ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="Discover_Skipped_Steps">}\href{../../Magellan/html/Process.html#method-Discover_Skipped_Steps}{\code{Magellan::Process$Discover_Skipped_Steps()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="EncapsulateScreens">}\href{../../Magellan/html/Process.html#method-EncapsulateScreens}{\code{Magellan::Process$EncapsulateScreens()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="GetScreens_ui">}\href{../../Magellan/html/Process.html#method-GetScreens_ui}{\code{Magellan::Process$GetScreens_ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="Set_All_Reset">}\href{../../Magellan/html/Process.html#method-Set_All_Reset}{\code{Magellan::Process$Set_All_Reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="Set_All_Skipped">}\href{../../Magellan/html/Process.html#method-Set_All_Skipped}{\code{Magellan::Process$Set_All_Skipped()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="ToggleState_Screens">}\href{../../Magellan/html/Process.html#method-ToggleState_Screens}{\code{Magellan::Process$ToggleState_Screens()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="Process" data-id="ValidateCurrentPos">}\href{../../Magellan/html/Process.html#method-ValidateCurrentPos}{\code{Magellan::Process$ValidateCurrentPos()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Global_server"></a>}}
\if{latex}{\out{\hypertarget{method-Global_server}{}}}
\subsection{Method \code{Global_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Global_server(session, input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Description_server"></a>}}
\if{latex}{\out{\hypertarget{method-Description_server}{}}}
\subsection{Method \code{Description_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Description_server(input, output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Description_ui"></a>}}
\if{latex}{\out{\hypertarget{method-Description_ui}{}}}
\subsection{Method \code{Description_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Description_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normalize_server"></a>}}
\if{latex}{\out{\hypertarget{method-Normalize_server}{}}}
\subsection{Method \code{Normalize_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Normalize_server(input, output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normalize_ui"></a>}}
\if{latex}{\out{\hypertarget{method-Normalize_ui}{}}}
\subsection{Method \code{Normalize_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Normalize_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save_server"></a>}}
\if{latex}{\out{\hypertarget{method-Save_server}{}}}
\subsection{Method \code{Save_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Save_server(input, output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save_ui"></a>}}
\if{latex}{\out{\hypertarget{method-Save_ui}{}}}
\subsection{Method \code{Save_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$Save_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protein_Normalization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
